跑位置直线

+++ code/inc/MasterCtrl.h
@@ -8,8 +8,8 @@ typedef struct _Ctrl {
 	uint64_t leftTime;	   // 腿的运行时钟，集中管理
 	uint64_t rightTime;	   //
 
-	uint64_t runTime;
-	uint64_t reachTime;
+	float runTime;
+	float reachTime;
 
 	Vector2f PosZero;
